separate to BaseClient and RedisClient classes in node
